Bypass pipenv for Docker installations.
There's no real reason to run a virtual environment inside of a Docker container, other than I had issues understanding pipenv.  Creating a Pipfile.lock freezes the dependency tree to a fixed set of files.  Checking that in, we can both promote a more standard build for users, as well as avoid using a venv inside of Docker.

To update the lockfile, we'd just simply run `pipenv lock` when we want to do a fresh update of our dependencies.  Or use the script I've written which does the same, but uses Docker as an intermediary, so you don't have to install pipenv on your local host.