Some docker compose tweaks.
Most importantly, fixing issues with Postgres in compose.
I wrote it to use the latest version of postgres, which is a terrible idea for postgres.  So I fixed that.  In the meantime, newer versions of postgres don't like the way I mounted volumes, so I changed that to match the new postgres.

Also, a better dev experience
By using the watch stanza, the compose stack can auto-update in development situations, and be faster in production-like situations.
To use watch for development, just start the stack with a `docker compose up --watch` command.